Huge commit. DEVFS have a first implementation
Still need to understand how to properly mount a superblock
in a mount point. Actually is a bit hacky and, once you are
in the "devfs" filesystem the pwd does now work properly because
we are not able to access to the mount point superblock.
(Maybe a reference is needed)

Also we've finally almost removed the ugly "dev_io" calls from
the virtual file system interface (is only present for ramdisk).