Two phase memory initialization
WIth a "huge" amount of RAM we need to add memory to the buddy system in
two phases.

In the first phase we add the low 4MB.
In the last phase we add the rest.

The last phase may require to split memory in two areas.

All of this tricky stuff just because the buddy system support
structures require an amount of memory proportional to the available
system memory.

The second phase can thus assume that:
- the page fault handler is up and running
- there are enough freed frames already registered in the buddy
allocator to accomodate the buddy metadata structs.